more cool stuff

+++ src/main/java/eu/zavadil/java/spring/common/paging/JsonPage.java
@@ -1,14 +1,12 @@
 package eu.zavadil.java.spring.common.paging;
 
-import org.springframework.data.domain.Page;
-
 import java.util.List;
 
 public interface JsonPage<T> {
 
 	List<T> getContent();
 
-	int getTotalItems();
+	long getTotalItems();
 
 	int getPageSize();
